Se configura perfil para admin y guest para que el administrador pueda ver todas las tareas

+++ Backend/routes/api.php
@@ -16,25 +16,29 @@ use Illuminate\Support\Facades\Route;
 | be assigned to the "api" middleware group. Make something great!
 |
 */
-
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
+Route::post('register', [AuthController::class, 'register']); 
 
-Route::post('register', [AuthController::class, 'register']); // ruta del endPoint para registrar un usuario
-
-Route::post('login', [AuthController::class, 'login']);// ruta del endpoint para validar si un usuario existe
+Route::post('login', [AuthController::class, 'login']); 
 
 Route::middleware('auth:api')->group(function () {
-    // Route::post('logout', [AuthController::class, 'logout']);
-    Route::get('profile', [AuthController::class, 'profile']);
 
+    Route::get('profile', [AuthController::class, 'profile']); // Obtener el perfil
+
+    // Solo el admin puede ver todas las tareas, los usuarios normales solo pueden ver sus mismas tareas
     Route::get('tasks', [TaskController::class, 'index']);
+
+    // Crear una tarea (todos los usuarios pueden crear tareas)
     Route::post('tasks', [TaskController::class, 'store']);
-    Route::put('tasks/{task}', [TaskController::class, 'update']);
-    Route::delete('tasks/{task}', [TaskController::class, 'destroy']);
+
+    // Solo el admin o el dueño de la tarea pueden actualizarla
+    Route::put('tasks/{task}', [TaskController::class, 'update'])->middleware('admin_or_owner');
+
+    // Solo el admin o el dueño de la tarea pueden eliminarla
+    Route::delete('tasks/{task}', [TaskController::class, 'destroy'])->middleware('admin_or_owner');
 });
 
-Route::middleware('auth:api')->post('/logout', [AuthController::class, 'logout']);
-
+Route::middleware('auth:api')->post('/logout', [AuthController::class, 'logout']); // Cerrar sesión
